change bandwidth measurement

+++ src/main/scala/adaptivecep/distributed/HostActorBase.scala
@@ -167,7 +167,7 @@ trait HostActorBase extends Actor with ActorLogging with RequiresMessageQueue[Bo
     case EndThroughPutMeasurement(instant, actual) =>
       val senderDiff = java.time.Duration.between(throughputStartMap(sender())._1, instant)
       val receiverDiff = java.time.Duration.between(throughputStartMap(sender())._2, clock.instant())
-      val bandwidth = (senderDiff.toMillis / receiverDiff.toMillis) * ((1000 / senderDiff.toMillis) * throughputMeasureMap(sender()))
+      val bandwidth = (senderDiff.toMillis.toDouble / receiverDiff.toMillis.toDouble) * ((1000 / senderDiff.toMillis) * throughputMeasureMap(sender()))
       println("(" + senderDiff.toMillis + "/" + receiverDiff.toMillis +") * (( 1000" +  "/" + senderDiff.toMillis + ") * " + throughputMeasureMap(sender()) + "))")
       println(bandwidth, actual)
       send(sender(), ThroughPutResponse(bandwidth.toInt))
